+++ tsetmc/database/__init__.py
@@ -2,6 +2,7 @@ from json import dump as _dump
 
 from tsetmc.instruments import _DB_PATH, Instrument as _Instrument, _L18S
 from tsetmc.market_watch import market_watch_init as _market_watch_init
+from tsetmc import _get_par_tree, _read_html
 
 
 # todo: add tests for this module
@@ -44,3 +45,9 @@ def update_db_using_market_watch() -> None:
         print(f'{diff} new entries were added.')
     else:
         print('No new entries were found.')
+
+
+def boards() -> dict[int, str]:
+    # Also, available at http://en.tsetmc.com/Loader.aspx?ParTree=121C1913.
+    iloc = _read_html(_get_par_tree('111C1913'), header=0)[0].iloc
+    return dict(zip(iloc[:, 0], iloc[:, 1]))