+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Unreleased
 ----------
-* new method: `Introduction.introduction`
-* new method: `Introduction.ombud_messages`
+* feat!: cs 69 and flow 3 were removed from offline database.
+* feat: `Introduction.introduction`
+* feat: `Introduction.ombud_messages`
+* feat: `database.boards`
 * fixed: `database.add_instrument_to_db`
-* Feat!: cs 69 and flow 3 were removed from offline database.